+++ Src/Private/ConvertTo-Pdf-WaterMark.ps1
@@ -46,7 +46,6 @@ function ConvertTo-Pdf-WaterMark {
     process {
         try {
             Write-Verbose -Message "Trying to convert $($ImageInput.Name) object to PDF format. Destination Path: $DestinationPath."
-            # & $ImageMagickPath -quality 100 $ImageInput.FullName $DestinationPath
             [Diagrammer.ConvertImageToPDF]::ConvertPngToPdf($ImageInput.FullName, $DestinationPath)
         } catch {
             Write-Verbose -Message "Unable to convert $($ImageInput.Name) object to PDF format."